added 'media/shader/FrameInfo.ubo'
project
  added 'ENGINE_GL_DEBUG_BUFFER_SKIP_DYNAMIC' and 'PT_BUFFER_DYNAMIC_CONDITION_FILTER' macros
    to filter per-frame log spam of updating uniform buffers

Engine
  updated shader to handle Uniform Block 'FrameInfo'
  now prints more OpenGL platform-specific info on startup
  'GL_DITHER' is now disabled by default

gl::Buffer
  added C-array-based constructor
  renamed 'GetData()' to 'GetDataRef()'
  added 'BindBufferBase()'

gl::GlWrapper
  added 'BindBufferBase()'
  added 'GetActiveUniformsiv()'
  added 'GetUniformBlockIndex()'
  added 'GetUniformIndices()'
  added 'UniformBlockBinding()'

DrawingControl / DrawingManager
  added 'FrameInfo' Uniform Block handling

SystemControl / SystemManager
  refactored 'GetMaximumUniformBlockBindingPoints()' (now returns 'size_t' not 'int')
  initialization now queries GL for the num of available binding points

'StandardShaderProgram'
  removed junk data
  removed all data handling that was moved into uniform 'FrameInfo'
  added 'LinkUniformBlockFrameInfo()'

+++ include/engine/service/SystemControl.h
@@ -31,7 +31,7 @@ public:
     virtual std::string GetStrVRAMAvailable() const = 0;
     virtual std::string GetStrVRAMEngineUsage() const = 0;
 
-    virtual int GetMaximumUniformBlockBindingPoints() const = 0;
+    virtual size_t GetMaximumUniformBlockBindingPoints() const = 0;
     virtual int GetMaximumUniformBlocksCombined() const = 0;
     virtual int GetMaximumUniformBlocksCompute() const = 0;
     virtual int GetMaximumUniformBlocksFragment() const = 0;
